requirements.txt
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==3.1.0
colorzero==1.1
distro==1.5.0
frozendict==2.3.7
gpiozero==1.6.2
groundlight==0.7.4
idna==2.10
numpy==1.19.5
pidng==4.0.9
piexif==1.1.3
Pillow==8.1.2
pydantic==1.10.7
python-apt==2.2.1
python-dateutil==2.8.2
python-prctl==1.7
pytz==2023.3
requests==2.28.2
RPi.GPIO==0.7.0
simplejpeg==1.6.4
six==1.16.0
spidev==3.5
ssh-import-id==5.10
toml==0.10.1
typing-extensions==4.5.0
urllib3==1.26.15
v4l2-python3==0.3.2
